Initially hardcode 'public' tooting
Also tweaked to allow \n char in feed format string.

Nb. Toots defaulted to 'private' which means only subscribers would see
them. Ultimately this is better, but for purposes such as outreach, is
not so useful.

This feature should be editable per-feed.
The database and front end will need further attention for this.